Telco customer churn
Importing the files and immediately deleting the column "customerID", because it includes unique values and we don't need it for a classification algorithm.  The next step is dropping the 0 values and preprocessing of the data set. This is done by converting the object data types to categories. With the categorical data, I used cat.codes encoder to help me encode the string values to integers. 
After this step the data is converted to ints and preprocessed. I set y to be my target, x being the data set. Scaling the data with MinMaxScaler and splitting it on Train Test Split.
Using Logistic Regression, Linear SVC, SVC and MLP Classifier. Found the best score with MLP classifier of 97%.